Maintain the cumulative cost when solving problems
-Implemented this for the bridge crossing problem. The cost (time taken)
is displayed after each move

-This feature will allow A* to be implemented

+++ AIUtils/State.h
@@ -8,7 +8,7 @@ namespace AI {
 	class State {
 		// A given state of the puzzle
 		public:
-			virtual int getStateCode() const = 0;
+			virtual unsigned long getStateCode() const = 0;
 			virtual void getActions(std::vector<Action*>& actions) const = 0;
 			virtual std::string describe() const = 0;
